requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
filelock==3.9.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
Pillow==9.3.0
requests==2.28.1
sympy==1.11.1
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
timm==0.3.2
